refactor : Redis에서 값을 저장하거나 조회할 때, Key에 Prefix를 붙여주도록 변경 (#60)
