+++ src/main/java/gdsc/binaryho/imhere/config/redis/RedisConfig.java
@@ -1,4 +1,4 @@
-package gdsc.binaryho.imhere.config;
+package gdsc.binaryho.imhere.config.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
