+++ src/main/java/gdsc/binaryho/imhere/config/redis/RedisKeyPrefixes.java
@@ -1,0 +1,7 @@
+package gdsc.binaryho.imhere.config.redis;
+
+public class RedisKeyPrefixes {
+
+    public static final String ATTENDANCE_NUMBER_KEY_PREFIX = "lecture_id$";
+    public static final String VERIFICATION_CODE_KEY_PREFIX = "email$";
+}
