+++ src/main/java/gdsc/binaryho/imhere/core/auth/infrastructure/VerificationCodeRedisRepository.java
@@ -1,5 +1,6 @@
 package gdsc.binaryho.imhere.core.auth.infrastructure;
 
+import gdsc.binaryho.imhere.config.redis.RedisKeyPrefixes;
 import gdsc.binaryho.imhere.core.auth.application.port.VerificationCodeRepository;
 import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
@@ -11,21 +12,21 @@ import org.springframework.stereotype.Repository;
 public class VerificationCodeRedisRepository implements VerificationCodeRepository {
 
     private static final Integer VERIFICATION_CODE_EXPIRE_TIME = 10;
+    private final String KEY_PREFIX = RedisKeyPrefixes.VERIFICATION_CODE_KEY_PREFIX;
     private final RedisTemplate<String, String> redisTemplate;
 
     @Override
     public String getByEmail(String email) {
-        if (email == null) {
-            return null;
-        }
-
-        return redisTemplate.opsForValue().get(email);
+        String queryKey = KEY_PREFIX + email;
+        return redisTemplate.opsForValue().get(queryKey);
     }
 
     @Override
     public void saveWithEmailAsKey(String email, String verificationCode) {
+        String saveKey = KEY_PREFIX + email;
+
         redisTemplate.opsForValue().set(
-            email,
+            saveKey,
             verificationCode,
             VERIFICATION_CODE_EXPIRE_TIME,
             TimeUnit.MINUTES
